alloc: switch PL_ARRAY to size_t length
Not that I anticipate us ever having arrays even remotely close to this
in size, but it's the more correct type to use in principle.

+++ src/pl_alloc.h
@@ -107,7 +107,7 @@ void pl_ref_deref(struct pl_ref **ref);
 
 // Helper functions for dealing with arrays
 
-#define PL_ARRAY(type) struct { type *elem; int num; }
+#define PL_ARRAY(type) struct { type *elem; size_t num; }
 
 #define PL_ARRAY_RESIZE(parent, arr, len)                                       \
     do {                                                                        \
